requirements.txt
numpy
numba
pyarrow
gcsfs
google-cloud-bigquery
google-cloud-bigquery-storage
